% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usdaplants-package.R
\docType{package}
\name{usdaplants-package}
\alias{usdaplants}
\alias{usdaplants-package}
\title{usdaplants: USDA PLANTS Database}
\description{
A convenience data package providing the USDA PLANTS database as tibbles. Includes plant taxonomy, growth characteristics, and state-level distribution data for over 32,000 plant species in the United States.
}
\author{
\strong{Maintainer}: Kieran Healy \email{kjhealy@gmail.com} (\href{https://orcid.org/0000-0001-9114-981X}{ORCID})

}
\keyword{internal}
